:bug: Fixed 2 issues in gridmaps
First, log odds to prob was not being calculated at all. So I implemented the algorithm for this.
Second, inv_sensor_model.m contained an error where the wrong index was being refrenced for the y value. This was used for updating the map; for both the free cells and the end points.